% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcALK.R
\name{borrowALKfromNeighbourRFAs}
\alias{borrowALKfromNeighbourRFAs}
\title{calculateALKModel}
\usage{
borrowALKfromNeighbourRFAs(RFA, species, year, quarter, dat, ALK,
  lengthDivision)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year which the ALKs are calculated.}

\item{quarter}{The quarter of the year which the ALKs are calculated.}

\item{ca}{Data with age information}
}
\value{
Returns a list with ALK for each trawl haul
}
\description{
calculateALKModel
}
