% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculateCPUEs.R
\name{calcmCPUEnorthSea}
\alias{calcmCPUEnorthSea}
\title{calcmCPUEnorthSea}
\usage{
calcmCPUEnorthSea(species, year, quarter, dat, ALKprocedure, B, dimCPUE,
  lengthDivision, useICESindexArea = FALSE)
}
\arguments{
\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{rfa}{information about size of RFA's.}

\item{data}{the datras-data needed.}

\item{column}{column in the mCPUE matrix to calculate mCPUE}
}
\value{
Returns the mCPUE per length class in the given statistical rectangle
}
\description{
Internal function for calculating mCPUE in whole North sea.
}
