% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcALK.R
\name{calculateALKHaulbased}
\alias{calculateALKHaulbased}
\title{calculateALKNew}
\usage{
calculateALKHaulbased(RFA, species, year, quarter, ca, hl, lengthDivision,
  dat = NULL)
}
\arguments{
\item{RFA}{Roundfish area number.}

\item{species}{The species of interest.}

\item{year}{The year which the ALKs are calculated.}

\item{quarter}{The quarter of the year which the ALKs are calculated.}

\item{ca}{The CA needed for calculating the ALKs.}

\item{hl}{The HL needed for calculating the ALKs (since there can be trawl hauls without age information).}

\item{dat}{All data, this is needed when filling the ALK using different RFAs. If dat==null, then ALK is not borrowed from other RFAs.}
}
\value{
Returns a list with ALK for each trawl haul
}
\description{
calculateALKNew
}
