% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatesCPUEWithUncertainty.R
\name{CPUEnorthSea}
\alias{CPUEnorthSea}
\title{calcCPUEwholeNorthSea}
\usage{
CPUEnorthSea(species, year, quarter, dat, bootstrapProcedure, B = 10,
  ALKprocedure = "", doBootstrap = TRUE, onlySimulate = FALSE,
  lengthDivision = 1:150, samplesWithinEachIntervall = NULL,
  nSimHauls = NULL, useICESindexArea = FALSE)
}
\value{
Returns the mCPUE per age class in the whole North Sea
}
\description{
calcCPUEwholeNorthSea
}
