% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatesCPUEWithUncertainty.R
\name{CPUErfa}
\alias{CPUErfa}
\title{CPUErfa}
\usage{
CPUErfa(RFA, species, year, quarter, dat, bootstrapProcedure = "datras",
  B = 10, ALKprocedure = "datras", lengthDivision,
  useICESindexArea = FALSE)
}
\arguments{
\item{RFA}{The roundfish area of interest.}

\item{species}{The species of interest.}

\item{year}{The year of interest.}

\item{quarter}{The quarter of interest.}

\item{bootstrapProcedure}{The bootstrap procedure ("simple", "stratisfied", ...)}

\item{B}{The number of simulations in the selected bootstrap procedure}

\item{ALKprocedure}{Either datras or haulbased}
}
\value{
Returns the mCPUE per age class in the given RFA with uncertainty
}
\description{
.
}
