% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractStatRecIndexArea}
\alias{extractStatRecIndexArea}
\title{extractStatRecAll}
\usage{
extractStatRecIndexArea(species)
}
\value{
return a vector with all the statistical rectangles in the index area
}
\description{
extractStatRecAll
}
