% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obsInHL}
\alias{obsInHL}
\title{obsInHL}
\usage{
obsInHL(species, hl_hh, id, length)
}
\arguments{
\item{species}{The species of interest.}

\item{hl_hh}{The data of interest.}

\item{id}{The id of interest.}

\item{length}{The id of interest.}

\item{year}{The year of interest.}
}
\value{
Returns the number of observed fish of a given length
}
\description{
obsInHL
}
