% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDatrasData.R
\name{sampleCA}
\alias{sampleCA}
\title{sampleCA}
\usage{
sampleCA(ca_hh, species, quarter, lengthDivision,
  samplesWithinEachIntervall, hl_hh)
}
\arguments{
\item{ca_hh}{The quarter of interest.}
}
\value{
Returns a modified data set of the data used for calculating the CPUE. The data is modified by removing
observations in a certain procedure.
}
\description{
.
}
