% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDatrasData.R
\name{sampleCAHaul}
\alias{sampleCAHaul}
\title{sampleCAHaul}
\usage{
sampleCAHaul(obsThisHaul, lengthDivision, samplesWithinEachIntervall,
  species, hl_hh)
}
\arguments{
\item{datDetailed}{The quarter of interest.}

\item{RFA}{The roundfish area of interest.}
}
\value{
Returns a modified data set of the data used for calculating the CPUE. The data is modified by removing
observations in a certain procedure.
}
\description{
.
}
